backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
python-dotenv==1.0.0
asyncpg==0.29.0
pydantic==2.5.0
typing-extensions>=4.8.0
aiosqlite==0.20.0
PyJWT==2.8.0
bcrypt==4.0.1
passlib==1.7.4
cohere==5.5.3